+++ ingredient/ingredient-service-bus-namespace/CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.55](https://github.com/HomecareHomebase/azure-bake/compare/v0.1.54...v0.1.55) (2019-06-19)
+
+**Note:** Version bump only for package @azbake/ingredient-service-bus-namespace
+
+
+
+
+
+## [0.1.54](https://github.com/HomecareHomebase/azure-bake/compare/v0.1.53...v0.1.54) (2019-06-17)
+
+**Note:** Version bump only for package @azbake/ingredient-service-bus-namespace
+
+
+
+
+
+## [0.1.53](https://github.com/HomecareHomebase/azure-bake/compare/v0.1.52...v0.1.53) (2019-06-17)
+
+**Note:** Version bump only for package @azbake/ingredient-service-bus-namespace
+
+
+
+
+
+## [0.1.52](https://github.com/HomecareHomebase/azure-bake/compare/v0.1.51...v0.1.52) (2019-06-11)
+
+
+### Bug Fixes
+
+* Helper functions for Service Bus namespace ingredient ([#64](https://github.com/HomecareHomebase/azure-bake/issues/64)) ([fc68cf8](https://github.com/HomecareHomebase/azure-bake/commit/fc68cf8))
+
+
+
+
+
 ## [0.1.49](https://github.com/HomecareHomebase/azure-bake/compare/v0.1.48...v0.1.49) (2019-05-21)
 
 **Note:** Version bump only for package @azbake/ingredient-service-bus-namespace