+++ ingredient/ingredient-utils/CHANGELOG.md
@@ -3,6 +3,42 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.55](https://github.com/HomecareHomebase/azure-bake/compare/v0.1.54...v0.1.55) (2019-06-19)
+
+
+### Bug Fixes
+
+* forcing a docker pull when mixing ([#78](https://github.com/HomecareHomebase/azure-bake/issues/78)) ([4cc3658](https://github.com/HomecareHomebase/azure-bake/commit/4cc3658))
+* forcing full update ([2dca56c](https://github.com/HomecareHomebase/azure-bake/commit/2dca56c))
+
+
+
+
+
+## [0.1.54](https://github.com/HomecareHomebase/azure-bake/compare/v0.1.53...v0.1.54) (2019-06-17)
+
+**Note:** Version bump only for package @azbake/ingredient-utils
+
+
+
+
+
+## [0.1.53](https://github.com/HomecareHomebase/azure-bake/compare/v0.1.52...v0.1.53) (2019-06-17)
+
+**Note:** Version bump only for package @azbake/ingredient-utils
+
+
+
+
+
+## [0.1.52](https://github.com/HomecareHomebase/azure-bake/compare/v0.1.51...v0.1.52) (2019-06-11)
+
+**Note:** Version bump only for package @azbake/ingredient-utils
+
+
+
+
+
 ## [0.1.47](https://github.com/HomecareHomebase/azure-bake/compare/v0.1.46...v0.1.47) (2019-05-15)